Finding memory leaks and fixing bug
Running valgrind exposed numerous failures to free memory appropriately. These are all now fixed.

In addition it appears when initializing the dictionary table for the first time, I was incrementing
the table pointer which meant it was pointing at the last bucket instead of first. What I cannot
understand is why the rest of the code worked at all once this was the case. Merits understanding.